Rename IMU files
